backend/feedback/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "feedback"
version = "1.0.0"
requires-python = ">=3.12, <3.14"
authors = [{ name = "Andrew", email = "hello.andrew.yan@outlook.com" }]

dependencies = [
  "ollama==0.6.1",
  "spacy==3.8.11",
  "en-core-web-lg",
  "faiss-cpu==1.13.2",
]

[tool.uv.sources]
en-core-web-lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl" }